fix problem with koin

+++ app/src/main/java/com/example/eventsapp/domain/model/EventModel.kt
@@ -1,7 +1,7 @@
 package com.example.eventsapp.domain.model
 
 class EventModel (
-    val date: Int,
+    val date: Long,
     val description: String,
     val image: String,
     val price: Float,
